Version 2.0 converts each PWM channel to use an H-bridge.
This allows the controller to be able to drive a 3V brushless motor.  The
M0 processor is capable of generating the timing for the motors, though at
this time I don't know if it works with 6 GPIO pins or only with three.

Six was chosen so that a PWMs can be set to a high impedence state.  This
is useful in the case where multiple brushed motors are attached between either
+3v or ground and the PWM pin, though the alternative is to just pull the PWM
output to the same level.  That might be version 2.1 later.

Signed-off-by: Aaron Williams <aaron.w2@gmail.com>